Changed size_type to int in device_vector. Removed all the compiler warnings resulting due to this change.

+++ include/bolt/cl/detail/binary_search.inl
@@ -139,7 +139,7 @@ namespace bolt {
                     typeDefs,
                     binary_search_kernels,
                     compileOptions);
-                size_t totalThreads = globalThreads+residueGlobalThreads;
+                int totalThreads = globalThreads+residueGlobalThreads;
 
                 control::buffPointer result = ctl.acquireBuffer( sizeof( int ) * totalThreads,
                                                                 CL_MEM_ALLOC_HOST_PTR|CL_MEM_WRITE_ONLY );
@@ -221,7 +221,7 @@ namespace bolt {
                 bolt::cl::wait(ctl, l_mapEvent);
 
                 bool r = false;
-                for(size_t i=0; i<totalThreads; i++)
+                for(int i=0; i<totalThreads; i++)
                 {
                     if(h_result[i] == 1)
                     {
@@ -256,7 +256,7 @@ namespace bolt {
             {
 
                 typedef typename std::iterator_traits<ForwardIterator>::value_type Type;
-                size_t sz = (last - first);
+                int sz = static_cast<int>(last - first);
                 if (sz < 1)
                      return false;
 
@@ -312,7 +312,7 @@ namespace bolt {
                 bolt::cl::device_vector_tag )
             {
                 typedef typename std::iterator_traits<DVForwardIterator>::value_type iType;
-                size_t szElements = static_cast<size_t>(std::distance(first, last) );
+                int szElements = static_cast<int>(std::distance(first, last) );
                 bolt::cl::control::e_RunMode runMode = ctl.getForceRunMode(); // could be dynamic choice some day.
                 if(runMode == bolt::cl::control::Automatic)
                 {
@@ -362,7 +362,7 @@ namespace bolt {
             {
 
                 typedef typename std::iterator_traits<DVForwardIterator>::value_type iType;
-                size_t szElements = static_cast<size_t>(std::distance(first, last) );
+                int szElements = static_cast<int>(std::distance(first, last) );
                 if (szElements == 0)
                     return false;
